Added button style
