+++ src/app/page.tsx
@@ -1,27 +1,12 @@
 "use client";
 
+import { SIMPLE_STYLED_BUTTON_CLASS } from "@/const/style";
 import { getCustomerProfile, initAuth, initPayment } from "@/lib/frontend";
 import Image from "next/image";
-import { useEffect, useRef } from "react";
+
 
 export default function Home() {
-  const inputRef = useRef<HTMLInputElement | null>(null);
 
-  useEffect(() => {
-    const handleVisibilityChange = () => {
-      if (document.visibilityState === 'visible' && inputRef.current && inputRef.current === document.activeElement) {
-        inputRef.current.blur();  // Unfocus first
-        // inputRef.current.focus(); // Refocus to trigger the keyboard
-        console.log('here')
-      }
-    };
-
-    document.addEventListener('visibilitychange', handleVisibilityChange);
-
-    return () => {
-      document.removeEventListener('visibilitychange', handleVisibilityChange);
-    };
-  }, []);
 
   return (
     <div className="w-full h-screen flex flex-col justify-center items-center">
@@ -37,9 +22,8 @@ export default function Home() {
       </div>
       {/*Example of how to use the functions from the lib/frontend/index.ts*/}
       {/* uncomment the button below to enable exchange token button */}
-      <input ref={inputRef} name="Hello"></input>
       
-      <button
+      <button className={SIMPLE_STYLED_BUTTON_CLASS}
         onClick={() =>
           initAuth(
             (result) => {
@@ -57,7 +41,7 @@ export default function Home() {
         Exchange Token
       </button>
       {/* uncomment the button below to enable get customer profile button */}
-      <button
+      <button className={SIMPLE_STYLED_BUTTON_CLASS}
         onClick={async () => {
           const result = await getCustomerProfile(
             sessionStorage.getItem("accessToken") ?? ""
@@ -68,7 +52,7 @@ export default function Home() {
         Get Customer Profile
       </button>
       {/* uncomment the button below to enable payment button */}
-      <button
+      <button className={SIMPLE_STYLED_BUTTON_CLASS}
         onClick={() =>
           initPayment((errorCode, errorDescription) => {
             console.error(errorCode, errorDescription);
@@ -76,7 +60,7 @@ export default function Home() {
         }
       >
         Payment
-      </button>
+      </button> 
     </div>
   );
 }
